requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
docxtpl>=0.16.0
python-docx>=0.8.11
docx2pdf>=0.1.8

pillow>=9.0.0
ortools>=9.8.0
jinja2>=3.1.0
# Optional / platform-specific extras (documented in README):
# tkinterdnd2>=0.3.0  # drag&drop support (Windows)
# If you prefer the default Tkinter UI, `customtkinter` may be omitted
customtkinter>=5.2.0
